LayoutTests/http/tests/resourceLoadStatistics/only-accept-first-party-cookies-with-third-party-cookie-blocking-expected.txt

Tests the first-party-only cookie policy when third-party cookie blocking is enabled.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE


--------
Frame: '<!--frame1-->'
--------
Cookies are: test_cookie = 1

--------
Frame: '<!--frame2-->'
--------
Cookies are:

--------
Frame: '<!--frame3-->'
--------
Cookies are: test_cookie = 1

--------
Frame: '<!--frame4-->'
--------


--------
Frame: '<!--frame5-->'
--------
Cookies are:

--------
Frame: '<!--frame6-->'
--------
Cookies are: test_cookie = 1

--------
Frame: '<!--frame7-->'
--------
